requirements.txt
# Core Dependencies
Django>=5.1.7
djangorestframework>=3.15.2
django-cors-headers>=4.7.0
django-q2>=1.7.1
python-dotenv>=1.0.1

# Data & System
certifi==2025.1.31
charset-normalizer==3.4.1
filelock==3.17.0
idna==3.10
narwhals==1.30.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
patsy==1.0.1
plotly==6.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scipy==1.15.2
six==1.17.0
statsmodels==0.14.4
tzdata==2025.1
urllib3==2.3.0
yfinance>=0.2.54
